// packages/puppeteer-crawler/src/internals/puppeteer-launcher.d.ts
export interface PuppeteerLaunchContext extends BrowserLaunchContext<PuppeteerPlugin['launchOptions'], unknown> {
    /**
     *  `puppeteer.launch` [options](https://pptr.dev/api/puppeteer.launchoptions)
     */
    launchOptions?: PuppeteerPlugin['launchOptions'];
    /**
     * URL to a HTTP proxy server. It must define the port number,
     * and it may also contain proxy username and password.
     *
     * Example: `http://bob:pass123@proxy.example.com:1234`.
     */
    proxyUrl?: string;
    /**
     * If `true` and `executablePath` is not set,
     * Puppeteer will launch full Google Chrome browser available on the machine
     * rather than the bundled Chromium. The path to Chrome executable
     * is taken from the `CRAWLEE_CHROME_EXECUTABLE_PATH` environment variable if provided,
     * or defaults to the typical Google Chrome executable location specific for the operating system.
     * By default, this option is `false`.
     * @default false
     */
    useChrome?: boolean;
    /**
     * Already required module (`Object`). This enables usage of various Puppeteer
     * wrappers such as `puppeteer-extra`.
     *
     * Take caution, because it can cause all kinds of unexpected errors and weird behavior.
     * Crawlee is not tested with any other library besides `puppeteer` itself.
     */
    launcher?: unknown;
    /**
     * With this option selected, all pages will be opened in a new incognito browser context.
     * This means they will not share cookies nor cache and their resources will not be throttled by one another.
     * @default false
     */
    useIncognitoPages?: boolean;
}
/**
 * `PuppeteerLauncher` is based on the `BrowserLauncher`. It launches `puppeteer` browser instance.
 * @ignore
 */
export declare class PuppeteerLauncher extends BrowserLauncher<PuppeteerPlugin, unknown> {
    readonly configuration: Configuration;
    /**
     * @internal
     */
    protected static optionsShape: {
        launcher: z.ZodOptional<z.ZodCustom<import("@crawlee/types").Dictionary, import("@crawlee/types").Dictionary>>;
        proxyUrl: z.ZodOptional<z.ZodURL>;
        useChrome: z.ZodOptional<z.ZodBoolean>;
        useIncognitoPages: z.ZodOptional<z.ZodBoolean>;
        browserPerProxy: z.ZodOptional<z.ZodBoolean>;
        ignoreProxyCertificate: z.ZodOptional<z.ZodBoolean>;
        userDataDir: z.ZodOptional<z.ZodString>;
        launchOptions: z.ZodOptional<z.ZodCustom<import("@crawlee/types").Dictionary, import("@crawlee/types").Dictionary>>;
        userAgent: z.ZodOptional<z.ZodString>;
    };
    /** @internal */
    protected static optionsSchema: z.ZodObject<{
        launcher: z.ZodOptional<z.ZodCustom<import("@crawlee/types").Dictionary, import("@crawlee/types").Dictionary>>;
        proxyUrl: z.ZodOptional<z.ZodURL>;
        useChrome: z.ZodOptional<z.ZodBoolean>;
        useIncognitoPages: z.ZodOptional<z.ZodBoolean>;
        browserPerProxy: z.ZodOptional<z.ZodBoolean>;
        ignoreProxyCertificate: z.ZodOptional<z.ZodBoolean>;
        userDataDir: z.ZodOptional<z.ZodString>;
        launchOptions: z.ZodOptional<z.ZodCustom<import("@crawlee/types").Dictionary, import("@crawlee/types").Dictionary>>;
        userAgent: z.ZodOptional<z.ZodString>;
    }, z.core.$strict>;
    /**
     * All `PuppeteerLauncher` parameters are passed via an launchContext object.
     */
    constructor(launchContext?: PuppeteerLaunchContext, configuration?: Configuration);
    protected getDefaultHeadlessOption(): boolean;
}